functions/metadata-updater/requirements.txt
flask<3.0
requests
google-cloud-tasks
google-api-core
googleapis-common-protos
google-cloud-pubsub
